build_containers.sh: Add terminal COLUMNS/LINES environment variables so that docker's allocated tty will match the screen.
Docker doesn't seem to set the screen size correctly.  This is annoying
in, for instance, gdb since it thinks you have an 80 x 24 tty.  We can
pass in the current terminal size as environment variables and then it
works; resizing your console will also correct the issue but is somewhat
distasteful.