Dockerfile: Use Ubuntu 22.04 as the base image; add missing bzip2 library.
We update to Ubuntu 22.04 to get a newer version of gdb-multiarch (from
version 9 before to version 12 now).  This newer version supports
backtraces and debugging across interrupt handlers such as hard faults
and mem manage faults.  The older version of gdb wouldn't be able to
decode the stack past the interrupt entry, making it really hard to
figure out what line of code may have been doing an access violation.

Ubuntu 22.04 doesn't seem to have bzip2 by default which is needed for
decompressing some of the archives, so we also add it here.